Create vw to store view schema info
Support both MySQL and PG in this commit


Former-commit-id: 6162cc1147a3b19a7b71edaa981e648614e3c763

+++ api/table.go
@@ -39,8 +39,8 @@ type TableCreate struct {
 	// Standard fields
 	// Value is assigned from the jwt subject field passed by the client.
 	CreatorId int
-	CreatedTs int
-	UpdatedTs int
+	CreatedTs int64
+	UpdatedTs int64
 
 	// Related fields
 	DatabaseId int
